requirements.txt
flask==2.3.3
transformers==4.31.0
torch==2.0.1
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
werkzeug==2.3.7 